[#9] Create a navigation system to be used on every page
Create a navigation system which will include links to each of the
features of this website (currently just dashboard, login and
register).  The navigation system provides feedback to the user to
indicate which page they are on (with the current page being displayed
in with darker text and a blue bottom margin).

Closes #11.