[#2] Create shorten URL controller and add routing
Create a controller and add routing to handle the link shortener form
submission.  For now, this controller has a blank store() method (this
will be added in another subtask as it requires some logic to test if
the URL has been shortened before).

The shorten URL page is accessible at /u/new.

Closes #4.