[#2] Create URL model
Create the URL model to represent a shortened URL in the database.  The
URL model will store the full link as well as the slug of the shortened
URL.  It also includes the user who shortened the URL (where NULL
represents an unauthenticated user) and the total amount of visits the
shortened link has received.

Closes #5.