[#2] Handle link shortener view submission
When the submit button on the link shortener form is pressed, the data
will be handled by the URL controller and saved to the database. If the
URL has been shortened before, by the same user, the same URL will be
returned. This will prevent spam and allow for the retrieval of lost
URLs.

The generated short URL will be in the format of /u/xxxxxx where x is
in [a-zA-Z0-9].

Closes #6.